--- dexs/silentswap/index.ts
@@ -1,90 +1,0 @@
-import { CHAIN } from "../../helpers/chains";
-import type { FetchOptions, SimpleAdapter } from "../../adapters/types";
-
-/**
- * SilentSwap volume adapter.
- *
- * SilentSwap is a privacy-preserving cross-chain swap protocol. The on-chain
- * gateway contracts route per-order funds and intentionally are not exposed
- * to indexers — surfacing them would leak the user-level activity that the
- * protocol is designed to keep private.
- */
-
-const STATS_CONTRACT = "0x5894a9B8B342BDb1AD7570C3656eE406eE26f676"; // BNB mainnet, deployed 2026-05-25
-
-const ONE_DAY_SECONDS = 24 * 60 * 60;
-
-const volumeByDateAbi =
-    "function volumeByDate(uint64) view returns (" +
-    "uint128 cumulativeUsd, uint128 inflowUsd, uint128 outflowUsd, " +
-    "uint128 dailyUsd, uint64 transferCount, uint32 tokenCount, " +
-    "uint64 bnbBlock, uint64 publishedAt)";
-
-function yyyymmdd(ts: number): bigint {
-    const d = new Date(ts * 1000);
-    return BigInt(
-        d.getUTCFullYear() * 10_000 + (d.getUTCMonth() + 1) * 100 + d.getUTCDate(),
-    );
-}
-
-async function readInflowAtDate(api: FetchOptions["api"], date: bigint): Promise<bigint> {
-    const snap = await api.call({
-        target: STATS_CONTRACT,
-        abi: volumeByDateAbi,
-        params: [date.toString()],
-    });
-    // The decoded named tuple has `inflowUsd` at index 1; fall back to positional.
-    const inflow =
-        (snap as { inflowUsd?: string | bigint })?.inflowUsd ??
-        (snap as Array<string | bigint>)[1] ??
-        0n;
-    return BigInt(inflow);
-}
-
-const fetch = async (options: FetchOptions) => {
-    const { api, startTimestamp } = options;
-
-    const date = yyyymmdd(startTimestamp);
-    const prevDate = yyyymmdd(startTimestamp - ONE_DAY_SECONDS);
-
-    const [todayInflow6, prevInflow6] = await Promise.all([
-        readInflowAtDate(api, date),
-        readInflowAtDate(api, prevDate),
-    ]);
-
-    const dailyInflow6 =
-        todayInflow6 > prevInflow6 ? todayInflow6 - prevInflow6 : todayInflow6;
-
-    return {
-        dailyVolume: Number(dailyInflow6) / 1e6,
-    };
-};
-
-const adapter: SimpleAdapter = {
-    version: 2,
-    adapter: {
-        [CHAIN.BSC]: {
-            fetch,
-            start: "2026-05-25",
-            // The reporter stores per-date snapshots as monotonic current state
-            // (`volumeByDate[date]` is written once and never overwritten), so
-            // reading at the latest block returns correct historical day values
-            // without needing an archive node.
-            runAtCurrTime: true,
-            meta: {
-                methodology: {
-                    Volume:
-                        "Anonymized inflow series read from the SilentSwapStats " +
-                        "reporter contract on BNB Chain (" + STATS_CONTRACT + "). " +
-                        "dailyVolume = volumeByDate(date).inflowUsd − " +
-                        "volumeByDate(date-1).inflowUsd (clamped to ≥0). DefiLlama " +
-                        "accumulates the running total from this daily series. " +
-                        "Per-order gateway addresses are intentionally not exposed " +
-                        "to preserve user privacy.",
-                },
-            },
-        },
-    },
-};
-
-export default adapter;
